feat: move APIError to internal/http
Because APIError belongs to http client, makes sense to have in there.
